Move struct generation in own factory.
