+++ src/obrw_factory.c
@@ -1,0 +1,11 @@
+
+#include "obrw_factory.h"
+
+struct wallpaper obrwFactory_getWallpaperItem()
+{
+    struct wallpaper usedWallpaper;
+    usedWallpaper.arrayIndex = -1;
+    strcpy(usedWallpaper.name, "");
+
+    return usedWallpaper;
+}
